iceberg/tests: extract gmock matchers for reuse
